Swap multi-select patch with mouse support
I NEVER, not a single time used the functionality of the multi-select
patch (and even if I wanted to, dmenu has a simple implementation of it
by default). I wanted mouse support for how nice it is, and the two
patches are incompatible (I could probably make them work together, but
didn't feel like reading code).